Fix flaky test and coverage
This update fixes a test for TTS output by injecting a mocked
object that should never be garbage collected while the test is
running (something that would occasionally trip up the test runner
before). It also "improves" test coverage for the Android ASR
component by making it non-final and thus removing the need to
modify its byte code for testing.