PR feedback
This turns NLUService into a simple interface, representing
the asynchronous nature of classification by returning a
Future that the client can choose to immediately wait for
or inspect later. This removes the need for event listeners
and explicit event dispatch in NLU components.

The redundant type field for slots has also been removed.