Modify existing TensorFlow wrappers for NLU
This updates the TensorflowModel class to support multiple
inputs and outputs, which will be necessary to handle
intent and tag output for the NLU model.

Sizes of the input/output tensors are now read directly
from the tflite model metadata, and callers can specify
the position of a state input tensor if one exists.