Fix: Microphone sharing for Android ASR
This fixes an issue with the AndroidSpeechRecognizer class
that was resulting in no speech being sent to the ASR on
many devices.

It adds new input components capable of releasing the microphone
when ASR is active or avoiding capturing it altogether.

External management of the microphone is communicated via the speech
context; when Spokestack is not managing the microphone, pipeline
stages are not run.