Feature: Propagate partial ASR results to client
This feature passes partial speech recognition results available
from all current ASR providers along to the client for optional
display. It does this by populating the speech context's
transcript with the partial result and firing a new
PARTIAL_RECOGNITION event.

Nothing has changed about the final transcript; it is available
from the speech context when a RECOGNIZE event is fired.