Fix: workaround for Android ASR timeouts (#84)
This is a (hopefully temporary) workaround for a change we've
noticed in the Android ASR behavior. It's started sending
NO_MATCH errors for stretches of inactivity that used to result
in SPEECH_TIMEOUT errors. The former is treated as an actual error
in our library code, but the latter is a plain event, so having
them mismatched causes problems for clients. As a result we're
remapping NO_MATCH errors to timeouts.