fix: manage speech context better for Android ASR
The Android speech recognizer does not always send events in
the order we originally expected, so the former speech context
management was a bit too aggressive. The Android ASR no longer
deactivates the speech context from within its internal
RecognitionListener until it's sure the interaction is over (via
a final reognition event or an error), and it now handles altering
the context's managed flag instead of leaving that to the microphone
input handler.