change encoder reset fill values
The encoder models for the wakeword trigger/keyword recognizer
currently use a tanh activation, and all models we have seen have their
values for silence driven to the low side of the activation. We want
to preserve the lag benefits of filling the encoder buffer on reset,
but 0 is an unstable value for this activation, so we will now fill
with -1.