agc enhancements
These changes modify the automatic gain control to call webrtc's
virtualmic function, which is apparently required for measuring current
levels, in spite of the comments. The defaults for AGC have also been
tuned to match the levels reported on the iPhone in this new
configuration. The AGC hard limiter has also been disabled, because it
clips at the target level, which is undesirable when the target level
is below full scale.

The AGC has been modified to be placed by default before the noise
suppressor, which in my tests gave better noise filtering. This change
has been applied to all profiles and should be applied in applications
that don't use profiles.