Allow transcripts to be edited before NLU
This change allows automatic classifcation via the wrapper
class to be disabled if desired. It also incorporates a
feature from the React Native tray: the ability to specify
a function for editing ASR trasncripts before they are
passed through NLU for classification. This latter feature
is only active if automatic classification is also enabled.