breaking: Dialogue management API (#100)
Dialogue management is a subsystem designed to be connected to
NLU results. The dialogue manager processes NLU results and
translates them into system prompts and cues for changes in the
app via the use of a dialogue policy.

Clients may implement their own dialogue policies.

Spokestack's rule-based dialogue policy relies on an external
JSON configuration file to define the conversation's states
and system responses.

This represents a breaking change because it changes the package
the `TraceListener` interface is located in; this should not be a
problem for post-9.0 clients, as that interface is wrapped by
`SpokestackAdapter`.