feature: Include dialogue in Spokestack wrapper
This adds a dialogue manager to the Spokestack wrapper
class, disabling it by default so that a dialogue policy is not
required to use Spokestack.

It is a breaking change in the sense that it changes the
configuration key for dialogue policy configuration, but
no releases include dialogue management, so it's safe to roll
this into the upcoming major release.