breaking: Allow NLU module to release resources
This change makes the NLUService extend AutoCloseable, which forces a
`close()` method on all implementors. The existing service uses this to
release TensorFlow model and vocab resources.

On the NLU manager, `close()` has been duplicated as a convenience
method called `release()`, named for parallelism with the newly
added `prepare()` method, which is its inverse.

The NLU module was the last one to provide release/prepare
support, so adding it suggests a change to the Spokestack wrapper
API, removing the release/prepare methods used for the TTS module
and repurposing them to handle resources for both NLU and TTS.