breaking: TTS no longer responds to lifecycle
This change stops the built-in TTS player from observing and
responding to Android lifecycle events. An app with multiple
activities will want to decide for itself which events trigger
changes in the TTS system. The new default setting will allow
playback to continue across activity transitions if Spokestack
is established as a long-lived resource not tied to a single
activity/fragment.