fix: tighten task submission in TTS player
This change moves ExoPlayer preparation and content playing into
the same task when a new TTS URL is received, reducing the possibility
of a race condition in constructing the player when more than one TTS
response becomes available in quick succession.

Additionally, uses of `prepare` and `clear` on the internal media player
and media source have been consolidated to reduce churn.