feat: allow speech pipeline to be paused
This adds `pause` and `resume` methods to the speech pipeline
to allow listening to be temporarily suspsended without fully
releasing the pipeline's resources.

In turn, `prepare` and `release` in the `Spokestack` wrapper
have been removed, with `start` and `stop` controlling the
resources of all modules instead of just the speech pipeline.
`pause` and `resume` have been threaded through the wrapper
and are automatically called by TTS events.