fix: enforce ordering of TTS responses
Currently, TTS requests submitted in close proximity
can result in audio being delivered to the client in a
different order than the requests were submitted.

This change keeps the requests asynchronous (as they must
be for Android networking) while enforcing ordering for the
results by introducing a request queue in the TTS manager.